requirements.txt
azure-core
azure-identity
Faker
Flask
flask-cors
Flask-SQLAlchemy
langchain_community
langchain_openai
langchain-sqlserver
langgraph
pyodbc
python-dateutil
python-dotenv
pydantic
mcp
gunicorn
Werkzeug
azure-identity
azure-eventhub==5.11.5
ipykernel